260417_cloud_transformation_agent/backend/requirements.txt
fastapi~=0.115.0
uvicorn~=0.34.0
openai>=1.40.0
azure-identity>=1.13.0
pydantic>=2.0.0
pyyaml>=6.0.3
python-dotenv>=1.2.0
boto3>=1.34.0
